Switch to flutter DeviceOrientation enum
