+++ packages/camera/camera_platform_interface/lib/src/types/types.dart
@@ -7,4 +7,3 @@ export 'resolution_preset.dart';
 export 'camera_exception.dart';
 export 'flash_mode.dart';
 export 'exposure_mode.dart';
-export 'device_orientation.dart';
