+++ packages/camera/camera_platform_interface/test/events/device_event_test.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:camera_platform_interface/camera_platform_interface.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
@@ -10,40 +11,40 @@ void main() {
 
   group('DeviceOrientationChangedEvent tests', () {
     test('Constructor should initialize all properties', () {
-      final event = DeviceOrientationChangedEvent(DeviceOrientation.portrait);
+      final event = DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
 
-      expect(event.orientation, DeviceOrientation.portrait);
+      expect(event.orientation, DeviceOrientation.portraitUp);
     });
 
     test('fromJson should initialize all properties', () {
       final event = DeviceOrientationChangedEvent.fromJson(<String, dynamic>{
-        'orientation': 'portrait',
+        'orientation': 'portraitUp',
       });
 
-      expect(event.orientation, DeviceOrientation.portrait);
+      expect(event.orientation, DeviceOrientation.portraitUp);
     });
 
     test('toJson should return a map with all fields', () {
-      final event = DeviceOrientationChangedEvent(DeviceOrientation.portrait);
+      final event = DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
 
       final jsonMap = event.toJson();
 
       expect(jsonMap.length, 1);
-      expect(jsonMap['orientation'], 'portrait');
+      expect(jsonMap['orientation'], 'portraitUp');
     });
 
     test('equals should return true if objects are the same', () {
       final firstEvent =
-          DeviceOrientationChangedEvent(DeviceOrientation.portrait);
+          DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
       final secondEvent =
-          DeviceOrientationChangedEvent(DeviceOrientation.portrait);
+          DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
 
       expect(firstEvent == secondEvent, true);
     });
 
     test('equals should return false if orientation is different', () {
       final firstEvent =
-          DeviceOrientationChangedEvent(DeviceOrientation.portrait);
+          DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
       final secondEvent =
           DeviceOrientationChangedEvent(DeviceOrientation.landscapeLeft);
 
@@ -51,7 +52,7 @@ void main() {
     });
 
     test('hashCode should match hashCode of all properties', () {
-      final event = DeviceOrientationChangedEvent(DeviceOrientation.portrait);
+      final event = DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
       final expectedHashCode = event.orientation.hashCode;
 
       expect(event.hashCode, expectedHashCode);
