+++ packages/camera/camera_platform_interface/test/method_channel/method_channel_camera_test.dart
@@ -11,6 +11,7 @@ import 'package:camera_platform_interface/src/events/device_event.dart';
 import 'package:camera_platform_interface/src/method_channel/method_channel_camera.dart';
 import 'package:camera_platform_interface/src/utils/utils.dart';
 import 'package:flutter/services.dart' hide DeviceOrientation;
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -315,7 +316,8 @@ void main() {
         final streamQueue = StreamQueue(eventStream);
 
         // Emit test events
-        final event = DeviceOrientationChangedEvent(DeviceOrientation.portrait);
+        final event =
+            DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
         await camera.handleDeviceMethodCall(
             MethodCall('orientation_changed', event.toJson()));
         await camera.handleDeviceMethodCall(
